[refactor] API url 변경 필요
